Use cache for location data, removes local state
