+++ .gitignore
@@ -12,6 +12,7 @@ testing_notes
 /.cache
 
 # misc
+.vscode
 .DS_Store
 .env.local
 .env.development.local
